requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
langchain==0.2.15
langchain-community==0.2.14
langchain-core==0.2.36
langchain-google-genai==1.0.10